% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R, R/t-test.R, R/wilcox-test.R
\name{mean}
\alias{mean}
\alias{t_test}
\alias{wilcox_test}
\title{Inference for the Mean via Permutations}
\usage{
t_test(
  x,
  y = NULL,
  alternative = c("two_tail", "left_tail", "right_tail"),
  mu = 0,
  paired = FALSE,
  var_equal = FALSE,
  conf_level = 0.95,
  B = 1000
)

wilcox_test(
  x,
  y = NULL,
  alternative = c("two_tail", "left_tail", "right_tail"),
  mu = 0,
  paired = FALSE,
  conf_level = 0.95,
  B = 1000L
)
}
\arguments{
\item{x}{A numeric vector providing the first sample.}

\item{y}{A numeric vector providing the second sample.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two_tail"}, \code{"left_tail"} or \code{"right_tail"}. Defaults to
\code{"tow_tail"}.}

\item{mu}{A numeric value providing the value of the mean difference \code{mu_y - mu_x} under the null hypothesis. Defaults to \code{0}.}

\item{paired}{A boolean specifying whether the two samples are paired.
Defaults to \code{FALSE}. Unused at the moment since one-sample tests are not
yet available.}

\item{conf_level}{A numeric value specifying the confidence level for the
computation of a confidence interval for the mean difference. Defaults to
\verb{95\%}.}

\item{B}{An integer specifying the number of permutations to use. Defaults to
\code{1000L}.}
}
\value{
A list with class \code{"htest"} containing the following components:
\itemize{
\item \code{statistic}: the value of the test statistic computed from the original
observed samples.
\item \code{parameter}: the degrees of freedom of the null distribution of the test
statistic (not available for permutation tests).
\item \code{p.value}: the p-value for the test.
\item \code{estimate}: the estimated mean or mean difference \code{mu_y - mu_x} depending
on whether it was a one-sample test, a paired two-sample test or an
independent two-sample test.
\item \code{conf.int}: a confidence interval for the mean of for the mean difference
appropriate to the specified alternative hypothesis.
\item \code{null.value}: the specified hypothesized value of the mean or mean
difference depending on whether it was a one-sample test or a two-sample
test.
\item \code{stderr}: an estimate of the standard deviation of the estimator of the
mean or mean difference.
\item \code{alternative}: a string describing the alternative hypothesis.
\item \code{method}: a string indicating what type of test was performed.
\item \code{data.name}: a string giving the name(s) of the data.
}
}
\description{
Performs statistical inference tasks (point estimation, confidence interval,
tests) for the mean of the distributions of scalar-valued data.
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
z <- rnorm(10, mean = 3)
t1 <- t_test(x, y, B = 100)
t2 <- t_test(x, z, B = 100)
w1 <- wilcox_test(x, y, B = 100)
w2 <- wilcox_test(x, z, B = 100)
}
