% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-test.R
\name{t_test}
\alias{t_test}
\title{Student Permutation Test}
\usage{
t_test(
  x,
  y = NULL,
  alternative = c("two_tail", "left_tail", "right_tail"),
  mu = 0,
  paired = FALSE,
  var_equal = FALSE,
  conf_level = 0.95,
  B = 1000
)
}
\arguments{
\item{x}{A numeric vector providing the first sample.}

\item{y}{A numeric vector providing the second sample.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two_tail"}, \code{"left_tail"} or \code{"right_tail"}. Defaults to
\code{"tow_tail"}.}

\item{mu}{A numeric value providing the value of the mean difference under
the null hypothesis. Defaults to \code{0}.}

\item{paired}{A boolean specifying whether the two samples are paired.
Defaults to \code{FALSE}. Unused at the moment since one-sample tests are not
yet available.}

\item{var_equal}{A boolean specifying whether the variances of the two
distributions are equal or not. Defaults to \code{FALSE}.}

\item{conf_level}{A numeric value specifying the confidence level for the
computation of a confidence interval for the mean difference.}

\item{B}{An integer specifying the number of permutations to use. Defaults to
\code{1000L}.}
}
\value{
A list with class \code{"htest"} containing the following components:
\itemize{
\item \code{statistic}:
\item \code{parameter}:
\item \code{p.value}:
\item \code{estimate}:
\item \code{conf.int}:
\item \code{null.value}:
\item \code{stderr}:
\item \code{alternative}:
\item \code{method}:
\item \code{data.name}:
}
}
\description{
Performs a two-sample permutation test based on either Student or Welch
statistic on scalar-valued data. The parameter of interest if the difference
between the means of the two distributions.
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
z <- rnorm(10, mean = 3)
t1 <- t_test(x, y, B = 100)
t2 <- t_test(x, z, B = 100)
}
